+++ package.json
@@ -13,6 +13,7 @@
     "axios": "^1.7.7",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
+    "react-loading": "^2.0.3",
     "react-query": "^3.39.3",
     "react-redux": "^9.1.2",
     "react-router-dom": "^6.26.1",
