+++ src/components/Breadcrumbs/Breadcrumbs.tsx
@@ -3,7 +3,7 @@ import { useLocation } from 'react-router-dom'
 export const Breadcrumbs = () => {
 
     const location = useLocation()
-    // console.log(location)
+
     return (
         <div className='container'>
             <span>breadcrumbs</span>
