+++ src/components/Loader/index.js
@@ -1,0 +1,1 @@
+export { Loader } from "./Loader"
