+++ src/hooks/useSales.ts
@@ -1,0 +1,7 @@
+
+export const useSales = (newPrice: number | undefined, oldPrice: number | undefined): number | string => {
+    if( !newPrice || !oldPrice) {
+        return ''
+    }
+    return Math.floor( 100 - ((newPrice * 100) / oldPrice) )
+}
