+++ src/pages/CategoriesInnerPage.tsx
@@ -1,15 +1,12 @@
-import { useParams } from 'react-router-dom';
 import { CategoriesItems, Breadcrumbs } from "../components"
 
 export const CategoriesInnerPage = () => {
-
-    const { categoryID } = useParams()
 
     return (
         <main className="page inner-page">
             <Breadcrumbs />
             <div className='container'>
-                <CategoriesItems categoriesID={categoryID} />
+                <CategoriesItems />
             </div>
         </main>
     );
