+++ src/pages/ProductPage.tsx
@@ -5,10 +5,8 @@ export const ProductPage = () => {
     return (
         <main className="page inner-page">
             <Breadcrumbs />
-            <div className='container'>
-                <Product />
-                <PopularItems />
-            </div>
+            <Product />
+            <PopularItems />
         </main>
     );
 };
