--- src/types/Phone.ts
@@ -1,9 +1,0 @@
-export default interface IPhone {
-    id: number,
-    article: string,
-    name: string,
-    img: string,
-    label: string,
-    oldPrice: number,
-    newPrice: number
-}
