+++ src/types/User.ts
@@ -1,4 +1,4 @@
-export default interface IUser {
+export interface IUser {
     email: string,
     id: number,
     name: string,
