+++ src/types/index.js
@@ -1,0 +1,4 @@
+export * from "./Cart";
+export * from "./Categories";
+export * from "./Product";
+export * from "./User";
